// src/themes/types.ts
export interface IColorType {
  black: string;
  white: string;
  error: string;
  error10: string;
  black40: string;
  black60: string;
  snowFlake: string;
  success: string;
  warning: string;
  limoniteBrown: string;
  treeBarkGreen: string;
  metal: string;
  green: string;
  red: string;
  yellow: string;
  orange: string;
  skyBlue: string;
  vibrantBlue: string;
  purple: string;
  pink: string;
}